// frontend-admin/src/api/generated/model/backupScheduleConfigurationDto.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Kasse API
 * Registrierkasse API - RKSV uyumlu kasa sistemi
 * OpenAPI spec version: v1
 */
import type { BackupScheduleFrequency } from './backupScheduleFrequency';

export interface BackupScheduleConfigurationDto {
  /** @nullable */
  customCron?: string | null;
  /** @nullable */
  dayOfMonth?: number | null;
  /** @nullable */
  dayOfWeek?: number | null;
  frequency?: BackupScheduleFrequency;
  hourUtc?: number;
  minuteUtc?: number;
}